No-op: extract to separate concern
