+++ spec/modules/product/creation_limit_spec.rb
@@ -1,0 +1,36 @@
+# frozen_string_literal: true
+
+require "spec_helper"
+
+describe Product::CreationLimit do
+  let(:user) { create(:user) }
+
+  it "allows creating up to 10 products in 24 hours" do
+    create_list(:product, 9, user: user)
+    new_product = build(:product, user: user)
+    expect(new_product).to be_valid
+  end
+
+  it "prevents creating more than 10 products in 24 hours" do
+    create_list(:product, 10, user: user)
+    new_product = build(:product, user: user)
+    expect(new_product).not_to be_valid
+    expect(new_product.errors.full_messages).to include("Sorry, you can only create 10 products per day.")
+  end
+
+  it "allows different users to each create 10 products in 24 hours" do
+    user1 = create(:user)
+    user2 = create(:user)
+    create_list(:product, 10, user: user1)
+    new_product = build(:product, user: user2)
+    expect(new_product).to be_valid
+  end
+
+  it "allows creating products after 24 hours have passed" do
+    create_list(:product, 10, user: user, created_at: 25.hours.ago)
+    new_product = build(:product, user: user)
+    expect(new_product).to be_valid
+  end
+end
+
+
